+++ src/main/kotlin/icu/windea/pls/core/search/usages/ParadoxDefinitionUsagesSearcher.kt
@@ -20,7 +20,8 @@ class ParadoxDefinitionUsagesSearcher : QueryExecutorBase<PsiReference, Referenc
 		DumbService.getInstance(queryParameters.project).runReadActionInSmartMode {
 			val definitionInfo = target.definitionInfo ?: return@runReadActionInSmartMode
 			val name = definitionInfo.name
-			val useScope = target.useScope
+			//这里不能直接使用target.useScope，否则文件高亮会出现问题
+			val useScope = queryParameters.effectiveSearchScope
 			queryParameters.optimizer.searchWord(name, useScope, true, target)
 		}
 	}